REFACTOR move toelichtingen to separate component
